Impedir borrar objetos en uso

+++ models/actors.php
@@ -127,8 +127,27 @@ class Actor extends Template
         return Template::getAll("actors", "name, surnames");
     }
 
-    public static function deleteActor(int $id): bool
+    private static function canDeleteActor(int $id): bool
     {
+        self::initConnectionDb();
+        $sql = "
+            SELECT count(*) AS total
+            FROM Act
+            WHERE idActor = {$id}
+        ";
+        $query = self::$dbConnection->query($sql);
+        if (!$query) {
+            return false; // o lanza una excepción si prefieres
+        }
+        $row = $query->fetch_assoc();
+        return ((int)$row['total'] === 0);
+    }
+
+    public static function deleteActor(int $id): string
+    {
+        if(!Actor::canDeleteActor($id)) {
+            return "No se puede eliminar el actor porque está asociado a una o más series.";
+        }
         return Template::delete("actors", $id);
     }
 }
